+++ src/Symfony/Component/Filesystem/CHANGELOG.md
@@ -6,6 +6,11 @@ CHANGELOG
 
  * `Filesystem::dumpFile()` and `appendToFile()` don't accept arrays anymore
 
+4.4.0
+-----
+
+ * support for passing a `null` value to `Filesystem::isAbsolutePath()` is deprecated and will be removed in 5.0
+
 4.3.0
 -----