+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * made `Psr18Client` implement relevant PSR-17 factories
  * added `HttplugClient`
+ * added support for NTLM authentication
 
 4.3.0
 -----